requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
channels==2.1.5
channels-redis==2.3.1
constantly==15.1.0
daphne==2.2.2
dj-database-url==2.0.0
Django==2.1.3
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
psycopg2==2.9.6
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2018.7
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
typing-extensions==4.5.0
whitenoise==6.4.0
zope.interface==4.6.0
